feat: add lang recall, and ES6 modules
