+++ js/data.js
@@ -1,0 +1,74 @@
+const keyBoardData = [
+  { Backquote: ['`', '~', 'ё', 'Ё'] },
+  { Digit1: ['1', '!', '1', '!'] },
+  { Digit2: ['2', '@', '2', '"'] },
+  { Digit3: ['3', '#', '3', '№'] },
+  { Digit4: ['4', '$', '4', ';'] },
+  { Digit5: ['5', '%', '5', '%'] },
+  { Digit6: ['6', '^', '6', ':'] },
+  { Digit7: ['7', '&', '7', '?'] },
+  { Digit8: ['8', '*', '8', '*'] },
+  { Digit9: ['9', '(', '9', '('] },
+  { Digit0: ['0', ')', '0', ')'] },
+  { Minus: ['-', '_', '-', '_'] },
+  { Equal: ['=', '+', '=', '+'] },
+  { Backspace: ['Backspace', 'Backspace', 'Backspace', 'Backspace'] },
+
+  { Tab: ['Tab', 'Tab', 'Tab', 'Tab'] },
+  { KeyQ: ['q', 'q', 'й', 'й'] },
+  { KeyW: ['w', 'w', 'ц', 'ц'] },
+  { KeyE: ['e', 'e', 'у', 'у'] },
+  { KeyR: ['r', 'r', 'к', 'к'] },
+  { KeyT: ['t', 't', 'е', 'е'] },
+  { KeyY: ['y', 'y', 'н', 'н'] },
+  { KeyU: ['u', 'u', 'г', 'г'] },
+  { KeyI: ['i', 'i', 'ш', 'ш'] },
+  { KeyO: ['o', 'o', 'щ', 'щ'] },
+  { KeyP: ['p', 'p', 'з', 'з'] },
+  { BracketLeft: ['[', '{', 'х', 'х'] },
+  { BracketRight: [']', '}', 'ъ', 'ъ'] },
+  { Backslash: ['\\', '|', '\\', '/'] },
+  { Delete: ['Del', 'Del', 'Del', 'Del'] },
+
+  { CapsLock: ['CapsLock', 'CapsLock', 'CapsLock', 'CapsLock'] },
+  { KeyA: ['a', 'a', 'ф', 'ф'] },
+  { KeyS: ['s', 's', 'ы', 'ы'] },
+  { KeyD: ['d', 'd', 'в', 'в'] },
+  { KeyF: ['f', 'f', 'а', 'а'] },
+  { KeyG: ['g', 'g', 'п', 'п'] },
+  { KeyH: ['h', 'h', 'р', 'р'] },
+  { KeyJ: ['j', 'j', 'о', 'о'] },
+  { KeyK: ['k', 'k', 'л', 'л'] },
+  { KeyL: ['l', 'l', 'д', 'д'] },
+  { Semicolon: [';', ':', 'ж', 'ж'] },
+  { Quote: ['\'', '"', 'э', 'э'] },
+  { Enter: ['Enter', 'Enter', 'Enter', 'Enter'] },
+
+  { ShiftLeft: ['Shift', 'Shift', 'Shift', 'Shift'] },
+  { KeyZ: ['z', 'z', 'я', 'я'] },
+  { KeyX: ['x', 'x', 'ч', 'ч'] },
+  { KeyC: ['c', 'c', 'с', 'с'] },
+  { KeyV: ['v', 'v', 'м', 'м'] },
+  { KeyB: ['b', 'b', 'и', 'и'] },
+  { KeyN: ['n', 'n', 'т', 'т'] },
+  { KeyM: ['m', 'm', 'ь', 'ь'] },
+  { Comma: [',', '<', 'б', 'б'] },
+  { Period: ['.', '>', 'ю', 'ю'] },
+  { Slash: ['/', '?', '.', ','] },
+  { ArrowUp: ['\u25B2', '\u25B2', '\u25B2', '\u25B2'] },
+  { ShiftRight: ['Shift', 'Shift', 'Shift', 'Shift'] },
+
+  { ControlLeft: ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'] },
+  { MetaLeft: ['Win', 'Win', 'Win', 'Win'] },
+  { AltLeft: ['Alt', 'Alt', 'Alt', 'Alt'] },
+  { Space: [' ', ' ', ' ', ' '] },
+  { AltRight: ['Alt', 'Alt', 'Alt', 'Alt'] },
+  { ArrowLeft: ['\u25C2', '\u25C2', '\u25C2', '\u25C2'] },
+  { ArrowDown: ['\u25BE', '\u25BE', '\u25BE', '\u25BE'] },
+  { ArrowRight: ['\u25B8', '\u25B8', '\u25B8', '\u25B8'] },
+  { ControlRight: ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'] },
+];
+
+const body = document.querySelector('.body');
+
+export { keyBoardData, body };
